# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Smart Bill Manager API
  version: 1.0.0-m4
  description: Clean Slate M4 本地 API。所有租户资源均从服务端会话解析 TenantContext。
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /members/{user_id}:
    get:
      operationId: getMember
      tags: [accounts]
      description: 仅当前租户管理员按 ID 核对成员，冲突恢复不依赖其仍在当前分页。
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: 当前成员版本
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Member" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: changeMember
      tags: [accounts]
      description: 仅改变本租户角色/状态并撤销其旧会话；不修改全局账号。
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MemberChange" }
      responses:
        "200":
          description: 变更后的成员，版本递增
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Member" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /invitations/check:
    post:
      operationId: checkInvitation
      tags: [accounts]
      security: []
      description: 代码仅在请求体传输；无效、过期、撤销或已用统一拒绝。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvitationCode" }
      responses:
        "200":
          description: 持有代码者核对工作区与身份模式
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationView" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sbm_session
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 32
        maxLength: 256
...
  responses:
    BadRequest:
      description: 输入不合法
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: 未认证
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: 权限或 CSRF 校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: 资源不存在或对当前租户不可见
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: 幂等、版本、重复或状态冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Member:
      type: object
      additionalProperties: false
      required: [user_id, email, display_name, role, status, version, created_at]
      properties:
        user_id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        status: { type: string, enum: [active, suspended] }
        version: { type: integer, minimum: 1 }
        created_at: { type: string, format: date-time }
...
    MemberChange:
      type: object
      additionalProperties: false
      required: [role, status, expected_version, reason]
      properties:
        role: { $ref: "#/components/schemas/Role" }
        status: { type: string, enum: [active, suspended] }
        expected_version: { type: integer, minimum: 1 }
        reason: { type: string, minLength: 1, maxLength: 500 }
...
    InvitationCode:
      type: object
      additionalProperties: false
      required: [code]
      properties:
        code: { type: string, minLength: 43, maxLength: 43, writeOnly: true }
    InvitationView:
      type: object
      additionalProperties: false
      required: [tenant_name, email, role, existing_account, expires_at]
      properties:
        tenant_name: { type: string }
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/Role" }
        existing_account: { type: boolean }
        expires_at: { type: string, format: date-time }
...
    Error:
      $ref: ../schemas/error.schema.json
...
    Role:
      type: string
      enum: [owner, finance, reviewer, viewer]
...